move BraveFeatureList class to base package
