+++ android/brave_java_sources.gni
@@ -16,7 +16,6 @@ brave_java_sources = [
   "../../brave/android/java/org/chromium/chrome/browser/BraveAppHooks.java",
   "../../brave/android/java/org/chromium/chrome/browser/BraveApplicationImplBase.java",
   "../../brave/android/java/org/chromium/chrome/browser/BraveBadge.java",
-  "../../brave/android/java/org/chromium/chrome/browser/BraveFeatureList.java",
   "../../brave/android/java/org/chromium/chrome/browser/BraveHelper.java",
   "../../brave/android/java/org/chromium/chrome/browser/BraveIntentHandler.java",
   "../../brave/android/java/org/chromium/chrome/browser/BraveLaunchIntentDispatcher.java",
@@ -415,6 +414,7 @@ brave_browser_ui_android_toolbar_sources = [
 
 brave_java_base_sources = [
   "//brave/android/java/org/chromium/base/BraveCommandLineInitUtil.java",
+  "//brave/android/java/org/chromium/base/BraveFeatureList.java",
   "//brave/android/java/org/chromium/base/BraveReflectionUtil.java",
 ]
