+++ android/java/org/chromium/chrome/browser/notifications/retention/RetentionNotificationPublisher.java
@@ -14,11 +14,11 @@ import android.content.Intent;
 import android.net.Uri;
 
 import org.chromium.base.ApplicationStatus;
+import org.chromium.base.BraveFeatureList;
 import org.chromium.base.Log;
 import org.chromium.base.ThreadUtils;
 import org.chromium.base.task.AsyncTask;
 import org.chromium.chrome.browser.BraveAdsNativeHelper;
-import org.chromium.chrome.browser.BraveFeatureList;
 import org.chromium.chrome.browser.app.BraveActivity;
 import org.chromium.chrome.browser.brave_stats.BraveStatsUtil;
 import org.chromium.chrome.browser.flags.ChromeFeatureList;
