+++ android/java/org/chromium/chrome/browser/vpn/utils/InAppPurchaseWrapper.java
@@ -25,8 +25,8 @@ import com.android.billingclient.api.PurchasesUpdatedListener;
 import com.android.billingclient.api.SkuDetails;
 import com.android.billingclient.api.SkuDetailsParams;
 
+import org.chromium.base.BraveFeatureList;
 import org.chromium.chrome.R;
-import org.chromium.chrome.browser.BraveFeatureList;
 import org.chromium.chrome.browser.flags.ChromeFeatureList;
 import org.chromium.chrome.browser.vpn.utils.BraveVpnPrefUtils;
 import org.chromium.chrome.browser.vpn.utils.BraveVpnUtils;
